2010-04-25	Joel Sherrill <joel.sherrilL@OARcorp.com>
	* console/fbcons.c: Fix warnings.

+++ c/src/lib/libbsp/arm/csb337/console/fbcons.c
@@ -27,7 +27,7 @@
 static int     fbcons_first_open(int major, int minor, void *arg);
 static int     fbcons_last_close(int major, int minor, void *arg);
 static int     fbcons_read(int minor);
-static int     fbcons_write(int minor, const char *buf, int len);
+static ssize_t fbcons_write(int minor, const char *buf, size_t len);
 static void    fbcons_init(int minor);
 static void    fbcons_write_polled(int minor, char c);
 static int     fbcons_set_attributes(int minor, const struct termios *t);
@@ -91,7 +91,7 @@ static int fbcons_read(int minor)
  *
  * return 1 on success, -1 on error
  */
-static int fbcons_write(int minor, const char *buf, int len)
+static ssize_t fbcons_write(int minor, const char *buf, size_t len)
 {
   int i;
